# backend/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: OpenAPI definition
  version: v0
servers:
- url: https://localhost:8443
  description: Generated server url
paths:
  /api/users/{id}:
    get:
      tags:
      - user-rest-controller
      operationId: showUserProfile
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserNoImageDTO"
    put:
      tags:
      - user-rest-controller
      operationId: editUserProfile
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserNoImageDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
    delete:
      tags:
      - user-rest-controller
      operationId: deleteUser
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserNoImageDTO"
...
  /api/users/:
    get:
      tags:
      - user-rest-controller
      operationId: showUsers
      parameters:
      - name: pageable
        in: query
        required: true
        schema:
          $ref: "#/components/schemas/Pageable"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PagedModelUserNoImageDTO"
    post:
      tags:
      - user-rest-controller
      operationId: registerUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserNoImageDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserNoImageDTO"
...
  /api/users/{id}/image:
    get:
      tags:
      - user-rest-controller
      operationId: getProfileImage
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
...
components:
  schemas:
    CommentBasicDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        createdDate:
          type: string
          format: date
        text:
          type: string
    CourseBasicDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        description:
          type: string
        image:
          type: boolean
        numberOfUsers:
          type: integer
          format: int32
        tags:
          type: array
          items:
            type: string
    UserNoImageDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        email:
          type: string
        password:
          type: string
        courses:
          type: array
          items:
            $ref: "#/components/schemas/CourseBasicDTO"
        comments:
          type: array
          items:
            $ref: "#/components/schemas/CommentBasicDTO"
        roles:
          type: array
          items:
            type: string
...
    Pageable:
      type: object
      properties:
        page:
          type: integer
          format: int32
          minimum: 0
        size:
          type: integer
          format: int32
          minimum: 1
        sort:
          type: array
          items:
            type: string
    PageMetadata:
      type: object
      properties:
        size:
          type: integer
          format: int64
        number:
          type: integer
          format: int64
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int64
    PagedModelUserNoImageDTO:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/UserNoImageDTO"
        page:
          $ref: "#/components/schemas/PageMetadata"
...